requirements.txt
streamlit
llama-index
faiss-cpu
pypdf
python-docx
openai
